requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Twitter API
tweepy>=4.14.0

# Data generation for testing
faker>=18.0.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# NLP and Sentiment Analysis
textblob>=0.17.1
nltk>=3.8.0

# Optional for dashboards (uncomment if needed)
# streamlit>=1.24.0
# plotly>=5.17.0
